Refactor: Update CaseViewModel and ExhibitCatalogItem for type safety
This commit introduces changes to improve type safety and clarity in `CaseViewModel` and `ExhibitCatalogItem`.

Key changes:
-   **CaseViewModel.kt**:
    -   Explicitly typed local flow variables (`allegationsFlow`, `catalogFlow`, `caseExhibitsFlow`) for better readability and type inference.
    -   Provided explicit type arguments to the `combine` operator for `Flow`s of `Allegation`, `ExhibitCatalogItem`, and `Exhibit` to produce `List<DisplayExhibit>`.
    -   Renamed lambda parameters in the `combine` block for clarity (`allegations` to `currentAllegations`, `catalog` to `catalogItems`, `caseExhibits` to `caseExhibitsList`).
-   **ExhibitCatalogItem.kt**:
    -   Changed the type of `applicableAllegationIds` from `List<String>` to `List<Int>` to align with the expected data type for allegation IDs.

+++ app/src/main/java/com/hereliesaz/lexorcist/data/repository/ExhibitRepositoryImpl.kt
@@ -35,7 +35,7 @@ class ExhibitRepositoryImpl @Inject constructor(
                                 id = tokens[0],
                                 type = tokens[1],
                                 description = tokens[2],
-                                applicableAllegationIds = tokens[3].split(",").map { it.trim() }
+                                applicableAllegationIds = tokens[3].split(",").map { it.trim().toInt() }
                             )
                             exhibits.add(exhibit)
                         }
